Refactor api
Added API routers to simplify main file
