+++ api/src/models.py
@@ -1,0 +1,34 @@
+from sqlmodel import SQLModel, Field
+from pydantic import BaseModel
+from typing import Optional
+from datetime import date
+
+
+class UserAuth(BaseModel):
+    """
+    This is used for authenticating a user
+    """
+    username: str
+    password: str
+
+
+class User(SQLModel, table=True):
+    """
+    This is the schema for a user in the database. 
+
+    id: Generated by database to identify user
+    username: user inputed screen name
+    password: user password to authenticate their account
+    dob: user date of birth for ensuring age
+    credits: Quantity of credits owned by user
+    """
+    id: Optional[int] = Field(default=None, primary_key=True)
+    username: str
+    password: str
+    dob: date 
+    credits: Optional[int] = Field(default=0)
+
+class Game(SQLModel):
+    game_id: int
+    players: list
+
