+++ frontend/src/signup.html
@@ -29,7 +29,7 @@
             });
 
             // Send POST request to server
-            fetch('http://localhost:8000/register', {
+            fetch('http://localhost:8000/auth/register', {
                 method: 'POST',
                 headers: {
                     'Content-Type': 'application/json'
